+++ openunmix/transforms.py
@@ -1,0 +1,252 @@
+from typing import Optional
+
+import torch
+import torchaudio
+from asteroid_filterbanks.enc_dec import Encoder, Decoder
+from asteroid_filterbanks.transforms import to_torchaudio, from_torchaudio
+from asteroid_filterbanks import torch_stft_fb
+from torch import Tensor
+import torch.nn as nn
+
+
+class AsteroidSTFT(nn.Module):
+    def __init__(
+        self,
+        n_fft=4096,
+        n_hop=1024,
+        center=False,
+        sample_rate=44100.0
+    ):
+        super(AsteroidSTFT, self).__init__()
+
+        window = nn.Parameter(
+            torch.hann_window(n_fft),
+            requires_grad=False
+        )
+        fb = torch_stft_fb.TorchSTFTFB.from_torch_args(
+            n_fft=n_fft,
+            hop_length=n_hop,
+            win_length=n_fft,
+            window=window,
+            center=center,
+            sample_rate=sample_rate
+        )
+
+        self.enc = Encoder(fb)
+
+    def forward(self, x):
+        aux = self.enc(x)
+        return to_torchaudio(aux)
+
+
+class AsteroidISTFT(nn.Module):
+    """Multichannel Inverse-Short-Time-Fourier functional
+    wrapper for torch.istft to support batches
+
+    Args:
+        STFT (Tensor): complex stft of
+            shape (nb_samples, nb_channels, nb_bins, nb_frames, complex=2)
+            last axis is stacked real and imaginary
+        n_fft (int, optional): transform FFT size. Defaults to 4096.
+        n_hop (int, optional): transform hop size. Defaults to 1024.
+        window (callable, optional): window function
+        center (bool, optional): If True, the signals first window is
+            zero padded. Centering is required for a perfect
+            reconstruction of the signal. However, during training
+            of spectrogram models, it can safely turned off.
+            Defaults to `true`
+        length (int, optional): audio signal length to crop the signal
+
+    Returns:
+        x (Tensor): audio waveform of
+            shape (nb_samples, nb_channels, nb_timesteps)
+
+    """
+    def __init__(
+        self,
+        n_fft: int = 4096,
+        n_hop: int = 1024,
+        center: bool = False,
+        sample_rate: float = 44100.0
+    ):
+        super(AsteroidISTFT, self).__init__()
+
+        window = nn.Parameter(
+            torch.hann_window(n_fft),
+            requires_grad=False
+        )
+        fb = torch_stft_fb.TorchSTFTFB.from_torch_args(
+            n_fft=n_fft,
+            hop_length=n_hop,
+            win_length=n_fft,
+            window=window,
+            center=center,
+            sample_rate=sample_rate
+        )
+
+        self.dec = Decoder(fb)
+
+    def forward(self, X: Tensor, length: Optional[int] = None) -> Tensor:
+        aux = from_torchaudio(X)
+        return self.dec(aux, length=length)
+
+
+class TorchSTFT(nn.Module):
+    """Multichannel Short-Time-Fourier Forward transform
+    uses hard coded hann_window.
+    Args:
+        n_fft (int, optional): transform FFT size. Defaults to 4096.
+        n_hop (int, optional): transform hop size. Defaults to 1024.
+        center (bool, optional): If True, the signals first window is
+            zero padded. Centering is required for a perfect
+            reconstruction of the signal. However, during training
+            of spectrogram models, it can safely turned off.
+            Defaults to `true`
+    """
+    def __init__(
+        self,
+        n_fft=4096,
+        n_hop=1024,
+        center=False
+    ):
+        super(TorchSTFT, self).__init__()
+        self.window = nn.Parameter(
+            torch.hann_window(n_fft),
+            requires_grad=False
+        )
+        self.n_fft = n_fft
+        self.n_hop = n_hop
+        self.center = center
+
+    def forward(self, x: Tensor) -> Tensor:
+        """STFT forward path
+        Args:
+            x (Tensor): audio waveform of
+                shape (nb_samples, nb_channels, nb_timesteps)
+        Returns:
+            STFT (Tensor): complex stft of
+                shape (nb_samples, nb_channels, nb_bins, nb_frames, complex=2)
+                last axis is stacked real and imaginary
+        """
+
+        shape = x.size()
+        nb_samples, nb_channels, nb_timesteps = shape
+
+        # pack batch
+        x = x.view(-1, shape[-1])
+
+        stft_f = torch.stft(
+            x,
+            n_fft=self.n_fft,
+            hop_length=self.n_hop,
+            window=self.window,
+            center=self.center,
+            normalized=False,
+            onesided=True,
+            pad_mode='reflect'
+        )
+
+        # unpack batch
+        stft_f = stft_f.view(shape[:-1] + stft_f.shape[-3:])
+        return stft_f
+
+
+class TorchISTFT(nn.Module):
+    """Multichannel Inverse-Short-Time-Fourier functional
+    wrapper for torch.istft to support batches
+    Args:
+        STFT (Tensor): complex stft of
+            shape (nb_samples, nb_channels, nb_bins, nb_frames, complex=2)
+            last axis is stacked real and imaginary
+        n_fft (int, optional): transform FFT size. Defaults to 4096.
+        n_hop (int, optional): transform hop size. Defaults to 1024.
+        window (callable, optional): window function
+        center (bool, optional): If True, the signals first window is
+            zero padded. Centering is required for a perfect
+            reconstruction of the signal. However, during training
+            of spectrogram models, it can safely turned off.
+            Defaults to `true`
+        length (int, optional): audio signal length to crop the signal
+    Returns:
+        x (Tensor): audio waveform of
+            shape (nb_samples, nb_channels, nb_timesteps)
+    """
+    def __init__(
+        self,
+        n_fft: int = 4096,
+        n_hop: int = 1024,
+        center: bool = False,
+        sample_rate: float = 44100.0
+    ) -> None:
+        super(TorchISTFT, self).__init__()
+
+        self.n_fft = n_fft
+        self.n_hop = n_hop
+        self.center = center
+        self.sample_rate = sample_rate
+
+        self.window = nn.Parameter(
+            torch.hann_window(n_fft),
+            requires_grad=False
+        )
+
+    def forward(self, X: Tensor, length: Optional[int] = None) -> Tensor:
+        shape = X.size()
+        X = X.reshape(-1, shape[-3], shape[-2], shape[-1])
+
+        y = torch.istft(
+            X,
+            n_fft=self.n_fft,
+            hop_length=self.n_hop,
+            window=self.window,
+            center=self.center,
+            normalized=False,
+            onesided=True,
+            length=length
+        )
+
+        y = y.reshape(shape[:-3] + y.shape[-1:])
+
+        return y
+
+
+class ComplexNorm(nn.Module):
+    r"""Compute the norm of complex tensor input.
+
+    Extension of `torchaudio.functional.complex_norm` with mono
+
+    Args:
+        power (float): Power of the norm. (Default: `1.0`).
+        mono (bool): Downmix to single channel after applying power norm
+            to maximize
+    """
+
+    def __init__(
+        self,
+        power: float = 1.0,
+        mono: bool = False
+    ):
+        super(ComplexNorm, self).__init__()
+        self.power = power
+        self.mono = mono
+
+    def forward(self, spec: Tensor) -> Tensor:
+        """
+        Args:
+            spec: complex_tensor (Tensor): Tensor shape of
+                `(..., complex=2)`
+
+        Returns:
+            Tensor: Power/Mag of input
+                `(...,)`
+        """
+        # take the magnitude
+        spec = torchaudio.functional.complex_norm(
+            spec, power=self.power
+        )
+
+        # downmix in the mag domain to preserve energy
+        if self.mono:
+            spec = torch.mean(spec, 1, keepdim=True)
+
+        return spec
